requirements.txt
python-dotenv
discord
google-genai
google-generativeai